+++ django_grapesjs/forms/widgets.py
@@ -1,6 +1,7 @@
 from django import forms
-from django_grapesjs.settings import GRAPESJS_TEMPLATE
+from django_grapesjs import settings
 from django_grapesjs.utils import get_render_html_value
+from django_grapesjs.utils.get_source import get_grapejs_assets
 
 __all__ = (
     'GrapesJsWidget',
@@ -8,32 +9,19 @@ __all__ = (
 
 
 class GrapesJsWidget(forms.Textarea):
-    '''
+    """
     Textarea form widget with support grapesjs.
     This is widget base config grapesjs.
+    """
 
-    '''
-    template_name = GRAPESJS_TEMPLATE
+    template_name = settings.GRAPESJS_TEMPLATE
 
     class Media:
         css = {
-            'all': (
-                'css/django_grapesjs/grapes.min.css',
-                'css/django_grapesjs/grapesjs-preset-newsletter.css',
-                'css/django_grapesjs/grapesjs-preset-webpage.min.css',
-                'css/django_grapesjs/grapesjs-plugin-filestack.css',
-            )
+            'screen': get_grapejs_assets('css'),
+
         }
-        js = [
-            'js/django_grapesjs/feather-aviary-editor.js',
-            'js/django_grapesjs/grapes.js',
-            'js/django_grapesjs/grapesjs-aviary.min.js',
-            'js/django_grapesjs/grapesjs-preset-newsletter.min.js',
-            'js/django_grapesjs/grapesjs-preset-webpage.min.js',
-            'js/django_grapesjs/grapesjs-lory-slider.min.js',
-            'js/django_grapesjs/grapesjs-tabs.min.js',
-            'js/django_grapesjs/grapesjs-plugin-filestack.min.js',
-        ]
+        js = get_grapejs_assets('js'),
 
     def get_formatted_id_value(self, value_id):
         return value_id.replace('-', '_')
@@ -41,7 +29,9 @@ class GrapesJsWidget(forms.Textarea):
     def get_context(self, name, value, attrs):
         context = super().get_context(name, value, attrs)
 
-        context['widget']['attrs']['id'] = self.get_formatted_id_value(context['widget']['attrs']['id'])
+        context['widget']['attrs']['id'] = self.get_formatted_id_value(
+            context['widget']['attrs']['id']
+        )
         context['widget'].update({
             'get_render_html_value': get_render_html_value(
                 self.default_html, apply_django_tag=self.apply_django_tag