+++ example/app/admin.py
@@ -1,9 +1,10 @@
 from django.contrib import admin
-from app.models import ExampleModel
 from django_grapesjs.admin import GrapesJsAdminMixin
 
+from app.models import GrapesJSModel  # noqa: #E501
 
-@admin.register(ExampleModel)
+
+@admin.register(GrapesJSModel)
 class ExampleAdmin(GrapesJsAdminMixin, admin.ModelAdmin):
     pass